+++ backend/Mgtt.ECom/test/Mgtt.ECom.InfrastructureTest/Connectors/PayPalConnectorTest.cs
@@ -14,8 +14,24 @@ using Xunit;
 
 namespace Mgtt.ECom.InfrastructureTest.Connectors
 {
+    public class FakeHttpMessageHandler : HttpMessageHandler
+    {
+        private readonly Func<HttpRequestMessage, HttpResponseMessage> responseHandler;
+
+        public FakeHttpMessageHandler(Func<HttpRequestMessage, HttpResponseMessage> responseHandler)
+        {
+            this.responseHandler = responseHandler;
+        }
+
+        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            var response = this.responseHandler(request);
+            return Task.FromResult(response);
+        }
+    }
+
     public class PayPalConnectorTest
-{
+    {
     private readonly string orderId;
     
     private readonly IOptions<PayPalSettings> settings;
